Upgrade: Install/Add Storybook dependencies the default way
With this change we are respecting the user's setting related to how dependencies should be saved to the package.json. The default behaviour is adding a caret to dependencies when added to the package.json. People can override this behaviour by defining them in, for example, a `save-exact=true` setting in their .npmrc file.